캘리브레이션 - 패널 크기 조정 및 Alert 제거 처리/에러제한적용

+++ routes/config.py
@@ -167,7 +167,7 @@ def cali_setup():
 def set_saveref(data:CaliRef):
     try:
         redis_state.client.select(0)
-        refdata = {"U": int(data.U), "I": int(data.I), "In": int(data.In), "P": int(data.P), "Error": int(data.Error)}
+        refdata = {"U": int(data.U), "I": int(data.I), "In": int(data.In), "P": int(data.P), "Error": float(data.Error)}
         redis_state.client.hset("calibration", "ref", json.dumps(refdata))
         return {'passOK': '1'}
     except Exception as e:
